Add the add-organizations flag to environment command
